backend/requirements.txt
fastapi==0.115.7
uvicorn==0.30.6
pydantic==2.9.2
python-multipart==0.0.9
pillow==10.4.0
transformers==4.48.0
torch==2.4.1
accelerate==0.33.0
pytesseract==0.3.13
pdf2image==1.17.0
